feat: deleted tranche details
